fix: upgrade getIcons to uint64_t, add /condition talkaction, fix SQL quoting
- Change Condition::getIcons() return type from uint32_t to uint64_t
- Update all overrides (ConditionGeneric, ConditionDamage, ConditionSpeed,
  ConditionRooted, ConditionFeared, ConditionDrunk)
- Remove static_cast<uint32_t> from PlayerIcon_* usage
- Add /condition talkaction for testing (root/fear/agony/bleed)
- Fix roulette_plays uuid SQL quoting (add single quotes around %s)
- Update getClientIcons internal variable to uint64_t

+++ src/condition.cpp
@@ -383,7 +383,7 @@ bool Condition::isPersistent() const
 	return true;
 }
 
-uint32_t Condition::getIcons() const { return isBuff ? ICON_PARTY_BUFF : 0; }
+uint64_t Condition::getIcons() const { return isBuff ? ICON_PARTY_BUFF : 0; }
 
 bool Condition::updateCondition(const Condition* addCondition)
 {
@@ -421,9 +421,9 @@ void ConditionGeneric::addCondition(Creature*, const Condition* condition)
 	}
 }
 
-uint32_t ConditionGeneric::getIcons() const
+uint64_t ConditionGeneric::getIcons() const
 {
-	uint32_t icons = Condition::getIcons();
+	uint64_t icons = Condition::getIcons();
 
 	switch (conditionType) {
 		case CONDITION_MANASHIELD:
@@ -1644,9 +1644,9 @@ int32_t ConditionDamage::getTotalDamage() const
 	return std::abs(result);
 }
 
-uint32_t ConditionDamage::getIcons() const
+uint64_t ConditionDamage::getIcons() const
 {
-	uint32_t icons = Condition::getIcons();
+	uint64_t icons = Condition::getIcons();
 	switch (conditionType) {
 		case CONDITION_FIRE:
 			icons |= ICON_BURN;
@@ -1681,7 +1681,7 @@ uint32_t ConditionDamage::getIcons() const
 			break;
 
 		case CONDITION_AGONY:
-			icons |= static_cast<uint32_t>(PlayerIcon_Agony);
+			icons |= PlayerIcon_Agony;
 			break;
 
 		default:
@@ -1845,9 +1845,9 @@ void ConditionSpeed::addCondition(Creature* creature, const Condition* condition
 	}
 }
 
-uint32_t ConditionSpeed::getIcons() const
+uint64_t ConditionSpeed::getIcons() const
 {
-	uint32_t icons = Condition::getIcons();
+	uint64_t icons = Condition::getIcons();
 	switch (conditionType) {
 		case CONDITION_HASTE:
 			icons |= ICON_HASTE;
@@ -1894,10 +1894,10 @@ void ConditionRooted::endCondition(Creature* creature)
 	}
 }
 
-uint32_t ConditionRooted::getIcons() const
+uint64_t ConditionRooted::getIcons() const
 {
-	uint32_t icons = Condition::getIcons();
-	icons |= static_cast<uint32_t>(PlayerIcon_Rooted);
+	uint64_t icons = Condition::getIcons();
+	icons |= PlayerIcon_Rooted;
 	return icons;
 }
 
@@ -1951,10 +1951,10 @@ void ConditionFeared::addCondition(Creature*, const Condition* condition)
 	}
 }
 
-uint32_t ConditionFeared::getIcons() const
+uint64_t ConditionFeared::getIcons() const
 {
-	uint32_t icons = Condition::getIcons();
-	icons |= static_cast<uint32_t>(PlayerIcon_Feared);
+	uint64_t icons = Condition::getIcons();
+	icons |= PlayerIcon_Feared;
 	return icons;
 }
 
@@ -2334,7 +2334,7 @@ void ConditionDrunk::addCondition(Creature* creature, const Condition* condition
 
 void ConditionDrunk::endCondition(Creature* creature) { creature->setDrunkenness(0); }
 
-uint32_t ConditionDrunk::getIcons() const { return ICON_DRUNK; }
+uint64_t ConditionDrunk::getIcons() const { return ICON_DRUNK; }
 
 bool ConditionDrunk::setParam(ConditionParam_t param, int32_t value)
 {
